pyproject.toml
[tool.poetry]
name = "livraria"
version = "0.1.0"
description = ""
authors = ["Andre Oliveira <dev.andrefc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
